Reparación del enlace principal con la lista y textbox sueldo

+++ app/src/main/java/edu/ucne/registrodeprestamos/model/Ocupacion.kt
@@ -8,5 +8,5 @@ data class Ocupacion (
     @PrimaryKey(autoGenerate = true)
     val ocupacionId: Int = 0,
     val descripcion: String = "",
-    val sueldo: Int = 0
+    val sueldo: Float = 0.00f
 )
